array

+++ Sources/CounterSet.cpp
@@ -4,7 +4,6 @@
 */
 #include "CvGameCoreDLL.h"
 #include "CounterSet.h"
-#include "array.h"
 
 CounterSet::CounterSet() { init(8); }
 
@@ -20,8 +19,8 @@ void CounterSet::init(const int iNumCounters)
 {
 	iSize = safeCoercion(iNumCounters);
 	iNext = 0;
-	iArrCounters.init(static_cast<size_t>(iSize), 0);
-	iArrCurrentIDs.init(static_cast<size_t>(iSize), 0);
+	iArrCounters.init(iSize, 0);
+	iArrCurrentIDs.init(iSize, 0);
 }
 
 uint16_t CounterSet::getCount(const int ID)
